[Changes made by Monty...] "There is a potential problem with duplicate function names and form names.   Already we have designed a method of generating the searches in individual files where the filename directly indicates the function name within.   This convention will help us from accidentally creating duplicate function names, especially since some of you have a fondness for two character functions.   If a new search file is created with a function name already in use,  and is  attempted to be checked into CVS, an error will be generated due to the  duplication, and the author will know the function needs to be changed.
This convention could also be used to eliminate the same
problem with duplicate form names.   If the form name is
strictly based on the function name with a trailing 'f', then no
duplicate form names could be generated in the same way
as before..

Therefore I have modified the form names in some of the
searches to fit this convention.  While I had the files open
I removed the un-needed 'target="_blank"' declarations.

No functionality was changed.. Everything works as before."