Changes by Neel Doshi...
1.  Changed the name of the IsCommand function to multiIsCommand to avoid future name clashes.

2.  Changed the scope of some variables that do not need to be global.

3.  Added an alert for users who have their reuseBrowserWindowMode parameter set to 1 (in which case multi will spawn each search in the same window - not very useful).

4.  I'm not sure about this modification - In order to solve the problem of
multi not working in anything by IE, I added a 1.5 second delay between
searches only if the user's launchmode is not equal to zero.  This seems to
work in my version of Opera, but I have been unable to test this in other
browsers.  This may also be a sub-optimal fix.  Any feedback would be
appreciated.

5.  Changed the category to "Search the Web"