1.  search.css - I added a new file called search.css which contains all the styles used in the application (the bar, menu, help, and calendar).  Would someone mind changing the installation script to include this file, but not to overwrite it if it exists already? 2.  calendar.js - I modified the UI portion of the calendar to use the stylesheet instead of the variables formerly in search.htm. I also fixed a small bug where today does not highlight if today is Sunday.  This makes the calendar code much easier to read and modify. 3.  search.htm - I removed the styles and the calendar style variables, and added the stylesheet call. 4.  helpbox.js - This version of the help box uses the styles. 5.  categoryExpander.htc - For the stylesheet version of the help box. 6.  preferences.js - I added preferences for the helptime (requested by a few friends who are hooked on the deskbar).  I also added a set of preferences that can be used by search functions that wish to have some default behavior encoded in them. 7.  defer_tools.js - In this file I propose three changes that I think may make the process of writing complicated functions a bit easier.  Please let me know what you think:
parseArgs -
A.  Rewrote parse args to solve the problem of switches that were
never defined to be switches from being removed from the query.  This
caused problems with functions that use switches and expect a url
parameter (http://www.yahoo.com would become http:/.yahoo.com due to
the parsing).
B.  Added an optional third parameter to the function that will
disable the switch expansion (/b becomes /books).  A complicated
search that expects the '/' character in the query can use this to
avoid any problems.
C.  Added another output property which returns the switches in an
associative array.  In other words /foo:bar will be returned as
switch_val["foo"] with the value "bar".  I hope that this will make
it simpler to write functions with key/value pair switches.
The function should still be backwards compatible so existing
searches do not have to be modified.

isURL -
I added the function isURL which determines if a given string is an
http or https url or not.  If it is it will return the string, and if
it is not, it will return false.  If the string looks like a dns
name, the function will prefix the string with "http://" before
returning it.  I use this function in at least three different
functions so far, so I thought it would make sense for everyone to
share it.

convertStylesToInline -
I added a new function called convertStylesToInline which is used to
shoehorn the stylesheet into a window.createPopup spawned popup.  I
added it here because it is used by both the calendar and the menu.

- Neel