[From: Ryan Edwards]
I've changed a couple regexps to work with the new batch of silly TLDs like .biz and .info that have been introduced within the last year as well as a couple of uncommon existing ones that aren't already included. With the following changes, DQSD will properly recognize a URL with these suffixes.