Added fix from Gregg Tavares [gregg at greggman dot com]:
"I found a bug and the fix.  The bug is as the file has
no tag specifying the character set used, different
computers will interpret the file differently.
Specifically, running it on a Japanese XP machine IE
assumes the file is encoded in Japanese which makes
some character combinations turn into Japanese which
has the effect of generating systax errors in the
script.

The solution is to specify the character set so I
added these lines

<HEAD>
<meta http-equiv="Content-Type" content="text/html;
charset=iso-8859-1">
</HEAD>

just after the <html> tag and all of a sudden it
started working.

-gregg"